pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "gbtd-infra"
version = "0.1.0"
description = "Global Bug Tracker Dataset collection infrastructure"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "alembic>=1.13",
  "httpx>=0.27",
  "tenacity>=9.0",
  "typer>=0.12",
  "pydantic>=2.7",
  "pydantic-settings>=2.3",
  "PyYAML>=6.0",
  "tomlkit>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "mypy>=1.10",
]

[project.scripts]
gbtd = "gbtd_infra.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["gbtd_infra*"]

[tool.ruff]
line-length = 110
